% Generated by roxygen2: do not edit by hand
% Please edit documentation in ./import_module.R
\name{import_module_help}
\alias{import_module_help}
\title{Open documentation for \code{import_module}}
\usage{
import_module_help(self = FALSE)
}
\arguments{
\item{self}{A logical value. If \code{self = FALSE} (default), show
documentation for \code{import_module}. If \code{self = TRUE}, show
docuemntation for this function.}
}
\description{
Open a rendered HTML page of \code{import_module} documentation.
See \url{https://github.com/bobaekang/r-module} for a detailed explanation.
}
